% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_simplify_dataset.R
\docType{data}
\name{internal_simplify_dataset}
\alias{internal_simplify_dataset}
\title{A list of expressions for wrangling the source CSV file into a
more efficient format for analysis.}
\format{An object of class \code{list} of length 96.}
\usage{
internal_simplify_dataset
}
\description{
There are four main stages within this process:
Changing non-optional booleans which were stored as factors into
booleans.
Moving booleans for not entered or not applicable when applied to
dates into the date format by storing as 0 or negative constants.
Setting up bitfields for groups of logical values - ie.
thrombolysis no-but reasons or thrombolysis complications.
}
\details{
With selective column import, we only want to apply transformations
applicable to the columns we have imported, so we select the
relevant expressions from this list by their name and apply them.
}
\keyword{datasets}
