% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record_transform_classes.R
\name{is_record_transform}
\alias{is_record_transform}
\alias{is.record_transform}
\title{Test if the object is a record_transform}
\usage{
is_record_transform(x)
}
\arguments{
\item{x}{An object}
}
\value{
`TRUE` if the object inherits from the
`record_transform` class.
}
\description{
This function returns `TRUE` for record_transforms
}
