% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssnap_column_operation.R
\name{is_ssnap_column_operation}
\alias{is_ssnap_column_operation}
\alias{is.ssnap_column_operation}
\alias{is.ssnap_column_substitution}
\alias{is.ssnap_column_derivation}
\alias{is.ssnap_column_lookup}
\title{Test if the object is an ssnap_column_operation}
\usage{
is_ssnap_column_operation(x)
}
\arguments{
\item{x}{An object}
}
\value{
`TRUE` if the object inherits from the `ssnap_column_operation`
class.
}
\description{
This function returns `TRUE` for cohort_definitions
}
