% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record_transform_classes.R
\name{is_tidy_csv_boolean}
\alias{is_tidy_csv_boolean}
\alias{is.tidy_csv_boolean}
\title{Test if the object is a record_transform}
\usage{
is_tidy_csv_boolean(x)
}
\arguments{
\item{x}{An object}
}
\value{
`TRUE` if the object inherits from the
`record_transform` class.
}
\description{
This function returns `TRUE` for record_transforms
}
