% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_csv.R
\name{read_audit_csv}
\alias{read_audit_csv}
\title{Load SSNAP data from CSV file into memory}
\usage{
read_audit_csv(filename, column_names = NULL,
  ignore_fake_team_data = TRUE)
}
\arguments{
\item{filename}{Filename of the CSV file to load.}

\item{column_names}{A character vector of the column names which
are required. If this is not specified, all columns are imported.
Where possible, columns should be specified as this makes the
analysis significantly quicker.}

\item{ignore_fake_team_data}{A boolean. By default (TRUE) any team
codes known to be fake or test team codes are skipped. If FALSE,
all teams are exported.}
}
\value{
A tibble containing the imported SSNAP data.
}
\description{
This function imports data from a csv file exported by
www.strokeaudit.org.
}
\details{
Most SSNAP data exporting all teams will include a number of 'test'
or 'fake' teams. By default, we ignore these teams. If however you
need to include these teams (for example to test some new code,)
you can set ignore_fake_team_data to FALSE.
}
