% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record_transform_classes.R
\name{record_transform}
\alias{record_transform}
\title{Record transform classes}
\usage{
record_transform()
}
\description{
This file contains a set of operations to take raw CSV (or in the
future SQL) data and convert it to record-level data which can be
used for aggregation and output.
}
\details{
The file defines a parent class, record_transform, and a
set of subclasses representing different subclasses, and
associated helper function for creating one or more of these
objects.

This line is a piece of 'dummy' definition to get a feel for
What the transform objects will look like.
}
