% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssnap_column_operation.R
\name{ssnap_column_operation}
\alias{ssnap_column_operation}
\title{SSNAP Column operators.}
\usage{
ssnap_column_operation(.vars)
}
\description{
This is hopefully an evolution of record_transform_classes.
}
