% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_csv_to_pt_level_datatable.R
\docType{data}
\name{ssnap_record_transforms}
\alias{ssnap_record_transforms}
\title{This file contains experimental code with list-defined instructions
for building a patient level datatable from raw CSV.}
\format{An object of class \code{list} of length 50.}
\usage{
ssnap_record_transforms
}
\description{
The idea is that we reduce the import to a set of scriptable
actions that the import function can select from and are more
maintainable / generic.
}
\details{
It is expected we'll start down this road with a hybrid of hard
coding and scriptable components.
The following line can be used to produce a data frame from the
lists; but lists (ie. a vector of objects of class transforms) are
probably the best way to create the data initially.

test_frame <- do.call(rbind.data.frame, test_transforms)
}
\keyword{datasets}
