% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_simplify_dataset.R
\docType{data}
\name{tpa_complications}
\alias{tpa_complications}
\title{Bit-field flags for recording complications from thrombolysing a
patient}
\format{An object of class \code{numeric} of length 4.}
\usage{
tpa_complications
}
\value{
a named numeric vector of thrombolysis reasons and their
representative value within the bitfield.
}
\description{
In the raw data these are represented by a non-exclusive series
of Booleans. To save space and improve processing time, rather
than store the tPA complications as a series of TRUE/FALSE values,
we can store it as a binary number - where 0 is FALSE and 1 is
TRUE. Each parameter is therefore given a value of 1^X, where X is
incremented from 0.
}
\details{
tpa_complications is a vector of named integer values describing
which binary bit in a number represents which thrombolysis reason
so rather than having numeric values scattered in code which may be
unclear or need commenting every time you can refer to a named
vector item.

This allows us to take a few mathematical shortcuts.

To find if a value is TRUE you want to do a logical AND against
the value.

To find everyone who had no 'no buts' you just look for 0;
everyone who had at least one no but has a value > 0.
}
\keyword{datasets}
