% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_simplify_dataset.R
\docType{data}
\name{tpa_no_but}
\alias{tpa_no_but}
\title{tpa_no_but is a vector of named integer values describing which
binary bit in a number represents which thrombolysis reason so
rather than having numeric values scattered in code which may be
unclear or need commenting every time you can refer to a named
vector item.}
\format{An object of class \code{numeric} of length 10.}
\usage{
tpa_no_but
}
\value{
a named numeric vector of thrombolysis reasons and their
representative value within the bitfield.
}
\description{
This allows us to take a few mathematical shortcuts.
}
\details{
To find if a value is TRUE you want to do a logical AND against
the value.

To find everyone who had no 'no buts' you just look for 0;
everyone who had at least one no but has a value > 0.
}
\keyword{datasets}
