% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsafe_convert_ethnicity_codes.R
\name{unsafe_convert_ethnicity_codes}
\alias{unsafe_convert_ethnicity_codes}
\title{Convert ethnicity codes to UK-wide descriptions}
\usage{
unsafe_convert_ethnicity_codes(data_table)
}
\arguments{
\item{data_table}{A tibble containing source data including
Country and S1 Ethnicity}
}
\value{
The tibble but with EthnicityCatUK added
}
\description{
Ethnicity is stored as a set of letter-based codes. Unfortunately
England and Northern Ireland have different codes. In order to use
the information, we therefore need to use the code and country
information (which we have to get by converting the team code into
more detailed info including country) and convert it to a common
set for all areas.
}
\section{Why is it unsafe}{


The code was written some time back. However we don't pass the
tibble in functions any more (we pass quos so we can pack
sets of quos together in a single mutate statement), and for
speed we don't convert team codes and add country data unless it's
essential. As we don't have an active use for this code we haven't
converted it yet. We would need to convert it to quos, and
work out how we handle needing Country (ideally users wanting
ethnicity shouldn't need to specify the team code conversion).
}

